+++ AutomatorServer/ios/AutomatorServerUITests/MaestroServer.swift
@@ -1,45 +1,69 @@
 import Telegraph
 
 struct OpenAppCommand: Codable {
-  var id: String
+  var appId: String
+}
+
+struct TapCommand: Codable {
+  var appId: String
+  var selector: Selector
+}
+
+struct DoubleTapCommand: Codable {
+  var appId: String
+  var selector: Selector
+}
+
+struct EnterTextBySelectorCommand: Codable {
+  var appId: String
+  var data: String
+  var selector: Selector
+}
+
+struct EnterTextByIndexCommand: Codable {
+  var appId: String
+  var data: String
+  var index: Int
+}
+
+struct PermissionCommand: Codable {
+  var code: String
+}
+
+struct Selector: Codable {
+  var text: String
 }
 
 class MaestroServer {
-  private let envPortKey = "MAESTRO_PORT"
+  private static let envPortKey = "MAESTRO_PORT"
 
   private let port: Int
 
   private let server: Server
 
-  private let startTime: String
-
   private let automation = MaestroAutomation()
 
   private let dispatchGroup = DispatchGroup()
 
-  var isRunning: Bool {
-    server.isRunning
-  }
+  private let decoder = JSONDecoder()
+
+  private var passedPort: Int? = {
+    guard let portStr = ProcessInfo.processInfo.environment[envPortKey] else {
+      Logger.shared.i("\(envPortKey) is null")
+      return nil
+    }
+
+    return Int(portStr)
+  }()
 
   init() throws {
-    guard let portStr = ProcessInfo.processInfo.environment[envPortKey] else {
-      throw MaestroError.generic("\(envPortKey) is null")
-    }
-    guard let port = Int(portStr) else {
-      throw MaestroError.generic("\(envPortKey)=\(portStr) is not an Int")
-    }
-    self.port = port
-
+    self.port = passedPort ?? 8081
     self.server = Server()
-
-    let dateFormatter = DateFormatter()
-    dateFormatter.dateFormat = "HH:mm:ss"
-    startTime = dateFormatter.string(from: Date())
   }
 
   func setUpRoutes() {
     server.route(.GET, "") { request in
-      HTTPResponse(content: "Hello from AutomatorServer on iOS!\nStarted on \(self.startTime)")
+      HTTPResponse(content: "Hello from AutomatorServer on iOS!")
     }
 
     server.route(.GET, "isRunning") { request in
@@ -57,21 +81,79 @@ class MaestroServer {
     }
 
     server.route(.POST, "openApp") { request in
-      let decoder = JSONDecoder()
       do {
-        let command = try decoder.decode(OpenAppCommand.self, from: request.body)
-        self.automation.openApp(command.id)
+        let command = try self.decoder.decode(OpenAppCommand.self, from: request.body)
+        self.automation.openApp(command.appId)
         return HTTPResponse(.ok)
       } catch let err {
         return HTTPResponse(.badRequest, headers: [:], error: err)
       }
+    }
+
+    server.route(.POST, "tap") { request in
+      do {
+        let command = try self.decoder.decode(TapCommand.self, from: request.body)
+        self.automation.tap(on: command.selector.text, inApp: command.appId)
+        return HTTPResponse(.ok)
+      } catch let err {
+        return HTTPResponse(.badRequest, headers: [:], error: err)
+      }
+    }
+
+    server.route(.POST, "doubleTap") { request in
+      do {
+        let command = try self.decoder.decode(DoubleTapCommand.self, from: request.body)
+        self.automation.doubleTap(on: command.selector.text, inApp: command.appId)
+        return HTTPResponse(.ok)
+      } catch let err {
+        return HTTPResponse(.badRequest, headers: [:], error: err)
+      }
+    }
+
+    server.route(.POST, "enterTextBySelector") { request in
+      do {
+        let command = try self.decoder.decode(EnterTextBySelectorCommand.self, from: request.body)
+        self.automation.enterText(command.data, by: command.selector.text, inApp: command.appId)
+        return HTTPResponse(.ok)
+      } catch let err {
+        return HTTPResponse(.badRequest, headers: [:], error: err)
+      }
+    }
+
+    server.route(.POST, "enterTextByIndex") { request in
+      do {
+        let command = try self.decoder.decode(EnterTextByIndexCommand.self, from: request.body)
+        self.automation.enterText(command.data, by: command.index, inApp: command.appId)
+        return HTTPResponse(.ok)
+      } catch let err {
+        return HTTPResponse(.badRequest, headers: [:], error: err)
+      }
+    }
+
+    server.route(.POST, "handlePermission") { request in
+      do {
+        let command = try self.decoder.decode(PermissionCommand.self, from: request.body)
+        self.automation.handlePermission(code: command.code)
+        return HTTPResponse(.ok)
+      } catch let err {
+        return HTTPResponse(.badRequest, headers: [:], error: err)
+      }
+    }
+
+    server.route(.POST, "selectFineLocation") { request in
+      self.automation.selectFineLocation()
+      return HTTPResponse(.ok)
+    }
+
+    server.route(.POST, "selectCoarseLocation") { request in
+      self.automation.selectCoarseLocation()
+      return HTTPResponse(.ok)
     }
   }
 
   func start() throws {
     Logger.shared.i("Starting server...")
     do {
-      server.route(.GET, "/") { (.ok, "Hello from AutomatorServer on iOS") }
       try server.start(port: 8081)
       setUpRoutes()
       dispatchGroup.enter()