+++ packages/maestro_cli/lib/src/features/drive/constants.dart
@@ -1,3 +1,5 @@
 const envHostKey = 'MAESTRO_HOST';
 const envPortKey = 'MAESTRO_PORT';
 const envVerboseKey = 'MAESTRO_VERBOSE';
+const envPackageNameKey = 'MAESTRO_APP_PACKAGE_NAME';
+const envBundleIdKey = 'MAESTRO_APP_BUNDLE_ID';